+++ packages/components-js/tests/vrt/specs/states/modal.vrt.ts
@@ -9,7 +9,7 @@ import {
   PSEUDO_STATES,
   setContentWithDesignSystem,
 } from '../../helpers';
-import { type Theme } from '@porsche-design-system/utilities-v2';
+import { type Theme } from '@porsche-design-system/styles';
 
 const component = 'modal';