+++ packages/components/src/components/tag/tag-styles.ts
@@ -1,36 +1,38 @@
+import {
+  borderRadiusSmall,
+  frostedGlassStyle,
+  spacingStaticXSmall,
+  textXSmallStyle,
+} from '@porsche-design-system/styles';
+import { type Theme } from '../../types';
 import { getCss, isHighContrastMode } from '../../utils';
 import {
   addImportantToEachRule,
   colorSchemeStyles,
   getFocusJssStyle,
   getInvertedThemedColors,
-  getResetInitialStylesForSlottedAnchor,
   getThemedColors,
   getTransition,
   hostHiddenStyles,
   hoverMediaQuery,
   prefersColorSchemeDarkMediaQuery,
-  type ThemedColors,
 } from '../../styles';
-import { borderRadiusSmall, textXSmallStyle } from '@porsche-design-system/utilities-v2';
-import type { TagColor, TagColorDeprecated } from './tag-utils';
-import { getThemedBackgroundHoverColor } from './tag-utils';
-import type { Theme } from '../../types';
 import { getThemedBackgroundColor } from './tag-shared-utils';
+import { getThemedBackgroundHoverColor, type TagColor, type TagColorDeprecated } from './tag-utils';
 
 export const getColors = (
-  themedColors: ThemedColors,
   tagColor: Exclude<TagColor, TagColorDeprecated>,
   theme: Theme
 ): {
-  primaryColor: string;
+  textColor: string;
   backgroundColor: string;
   backgroundHoverColor: string;
 } => {
+  const themedColors = getThemedColors(theme);
   const { primaryColor } = tagColor === 'primary' ? getInvertedThemedColors(theme) : themedColors;
 
   return {
-    primaryColor,
+    textColor: primaryColor,
     backgroundColor: getThemedBackgroundColor(tagColor, themedColors),
     backgroundHoverColor: getThemedBackgroundHoverColor(tagColor, themedColors, theme),
   };
@@ -38,45 +40,46 @@ export const getColors = (
 
 export const getComponentCss = (
   tagColor: Exclude<TagColor, TagColorDeprecated>,
+  compact: boolean,
   isFocusable: boolean,
+  hasIcon: boolean,
   theme: Theme
 ): string => {
-  const themedColors = getThemedColors(theme);
-  const themedColorsDark = getThemedColors('dark');
-  const { primaryColor, backgroundColor, backgroundHoverColor } = getColors(themedColors, tagColor, theme);
+  const { textColor, backgroundColor, backgroundHoverColor } = getColors(tagColor, theme);
   const {
-    primaryColor: primaryColorDark,
+    textColor: textColorDark,
     backgroundColor: backgroundColorDark,
     backgroundHoverColor: backgroundHoverColorDark,
-  } = getColors(themedColorsDark, tagColor, 'dark');
+  } = getColors(tagColor, 'dark');
+  const isBackgroundFrosted = tagColor === 'background-frosted';
 
   return getCss({
     '@global': {
       ':host': {
         display: 'inline-flex',
-        verticalAlign: 'top',
-        whiteSpace: 'nowrap',
+        verticalAlign: 'top', // TODO: should we set this CSS style at all?
+        whiteSpace: 'nowrap', // TODO: should either be exposed by a controlled CSS variable or a component prop or whitelist as supported custom styles
         ...addImportantToEachRule({
           ...colorSchemeStyles,
           ...hostHiddenStyles,
         }),
       },
       span: {
+        position: 'relative', // necessary as relative anchor to ensure click area of optional slotted focusable element is in sync
         display: 'flex',
         gap: '2px',
-        alignItems: 'center',
-        position: 'relative',
-        padding: '4px 9px',
+        padding: compact ? '1px 6px' : `${spacingStaticXSmall} 9px`,
         borderRadius: borderRadiusSmall,
-        background: backgroundColor,
-        color: primaryColor,
         font: textXSmallStyle.font,
+        color: textColor,
+        background: backgroundColor,
+        ...(isBackgroundFrosted && frostedGlassStyle),
         ...(isHighContrastMode && {
           outline: '1px solid transparent',
         }),
         ...prefersColorSchemeDarkMediaQuery(theme, {
+          color: textColorDark,
           background: backgroundColorDark,
-          color: primaryColorDark,
         }),
         ...(isFocusable &&
           hoverMediaQuery({
@@ -91,17 +94,11 @@ export const getComponentCss = (
       },
       '::slotted': addImportantToEachRule({
         '&(a),&(button)': {
-          ...getResetInitialStylesForSlottedAnchor,
-          display: 'inline',
-          position: 'static',
+          all: 'unset', // resets any ua-style + custom style set in light dom
           textDecoration: 'underline',
           cursor: 'pointer',
           font: 'inherit',
           color: 'inherit',
-          WebkitAppearance: 'none', // iOS safari
-          appearance: 'none',
-          border: 0,
-          textAlign: 'start',
         },
         '&(a)::before,&(button)::before': {
           content: '""',
@@ -116,14 +113,14 @@ export const getComponentCss = (
         },
       }),
     },
-    icon: {
-      marginInlineStart: '-2px', // compensate white space of svg icon and optimize visual alignment
-      alignSelf: 'flex-start',
-      ...(['neutral-contrast-high', 'primary'].includes(tagColor) && {
-        ...prefersColorSchemeDarkMediaQuery(theme, {
-          filter: 'invert(1)',
-        }),
-      }),
-    },
+    ...(hasIcon && {
+      icon: {
+        marginInlineStart: '-2px', // compensate white space of svg icon and optimize visual alignment
+        ...(!isHighContrastMode &&
+          tagColor === 'primary' && {
+            filter: 'invert(1)',
+          }),
+      },
+    }),
   });
 };