+++ packages/components/src/utils/color/changeColor.spec.ts
@@ -1,0 +1,19 @@
+import { lighten, darken } from './changeColor';
+
+it.each([
+  ['hsl(240 3% 26%/35%)', 'hsl(240 3% 36%/35%)'],
+  ['hsl(240 3% 99%/35%)', 'hsl(240 3% 100%/35%)'],
+  ['hsl(240 3% 26/35%)', 'hsl(240 3% 36/35%)'],
+  ['hsl(240 3% 99/35%)', 'hsl(240 3% 100/35%)'],
+])('should for lighten() hsl: %s return %s', (hsl, result) => {
+  expect(lighten(hsl)).toBe(result);
+});
+
+it.each([
+  ['hsl(240 3% 26%/35%)', 'hsl(240 3% 16%/35%)'],
+  ['hsl(240 3% 1%/35%)', 'hsl(240 3% 0%/35%)'],
+  ['hsl(240 3% 26/35%)', 'hsl(240 3% 16/35%)'],
+  ['hsl(240 3% 1/35%)', 'hsl(240 3% 0/35%)'],
+])('should for darken() hsl: %s return %s', (hsl, result) => {
+  expect(darken(hsl)).toBe(result);
+});